+++ src/mp-ci/index.ts
@@ -7,5 +7,10 @@ export {
   flattenSubPackages,
 } from './helper';
 export { MpCI } from './mp-ci';
+
+export { updateQQMpResultToSheetV0 } from './mp-sheet-v0';
+export { updateQQMpResultToSheet } from './mp-sheet';
+
+
 export { mpUploadAndReport, mpUploadAndReportByOptions } from './mp-upload-and-report';
 export { writeEnvAndPrivateKey, writeEnvAndPrivateKeyByOptions } from './write-env-and-private-key';