+++ src/mp/index.ts
@@ -1,1 +1,2 @@
+export { saveMpImage } from './save-mp-image';
 export { updateManager } from './update-manager';