+++ src/scheduler/types.ts
@@ -1,0 +1,5 @@
+export interface IRequest {
+  (): Promise<any>;
+  resolve?: (...args: any) => any;
+  reject?: (...args: any) => any;
+}